hal: Factor DMA channel code into dma::channel
Renames the `dma::DMA` type (originally named `ZeroDMA` in the Adafruit
code) into `dma::channel::Channel`, to reflect its role as a wrapper for
an individual channel.
