+++ hal/src/dma/controller/samd21.rs
@@ -1,8 +1,7 @@
 //! SAMD21 DMA controller (1 channel)
 
-use super::{ChannelId, ChannelMask};
-use crate::dma::Error;
-use crate::target_device::{dmac, DMAC, NVIC, Interrupt};
+use crate::dma::{channel, Error};
+use crate::target_device::{dmac, Interrupt, DMAC, NVIC};
 use core::ops::Deref;
 use vcell::VolatileCell;
 
@@ -24,7 +23,9 @@ pub(super) fn init_clock(clock_source: &mut ClockSource) {
 /// Enable DMAC IRQ
 pub(super) fn init_interrupts(nvic: &mut NVIC) {
     nvic.enable(Interrupt::DMAC);
-    unsafe { nvic.set_priority(Interrupt::DMAC, (1 << NVIC_PRIO_BITS) - 1); }
+    unsafe {
+        nvic.set_priority(Interrupt::DMAC, (1 << NVIC_PRIO_BITS) - 1);
+    }
 }
 
 /// Critical DMAC fields
@@ -33,7 +34,7 @@ pub(crate) struct ControllerCritical {
     dmac: DMAC,
 
     /// Bitmask of allocated channels
-    channel_mask: VolatileCell<ChannelMask>,
+    channel_mask: VolatileCell<channel::Mask>,
 }
 
 impl ControllerCritical {
@@ -51,29 +52,29 @@ impl ControllerCritical {
     }
 
     /// Read the current state of the channel mask
-    pub fn channel_mask(&self) -> ChannelMask {
+    pub fn channel_mask(&self) -> channel::Mask {
         self.channel_mask.get()
     }
 
     /// Read the current channel flags given a flag value
-    pub fn read_channel_id_and_flags(&self, flags: u8) -> (u8, ChannelId) {
+    pub fn read_channel_id_and_flags(&self, flags: u8) -> (u8, channel::Id) {
         (flags, 0)
     }
 
     /// Clear transfer error (TERR) flag
-    pub fn clear_transfer_error_flag(&self, channel_id: ChannelId) {
+    pub fn clear_transfer_error_flag(&self, channel_id: channel::Id) {
         debug_assert_eq!(channel_id, 0);
         self.dmac.chintflag.write(|reg| reg.terr().clear_bit());
     }
 
     /// Clear transfer complete (TCMPL) flag
-    pub fn clear_transfer_complete_flag(&self, channel_id: ChannelId) {
+    pub fn clear_transfer_complete_flag(&self, channel_id: channel::Id) {
         debug_assert_eq!(channel_id, 0);
         self.dmac.chintflag.write(|reg| reg.tcmpl().set_bit());
     }
 
     /// Clear channel suspend (SUSP) flag
-    pub fn clear_suspend_flag(&self, channel_id: ChannelId) {
+    pub fn clear_suspend_flag(&self, channel_id: channel::Id) {
         debug_assert_eq!(channel_id, 0);
         self.dmac.chintflag.write(|reg| reg.susp().set_bit());
     }
@@ -81,7 +82,7 @@ impl ControllerCritical {
     /// Configure default behaviors
     pub fn configure_default_behaviors(
         &self,
-        channel_id: ChannelId,
+        channel_id: channel::Id,
         peripheral_trigger: dmac::chctrlb::TRIGSRCW,
         trigger_action: dmac::chctrlb::TRIGACTW,
     ) {
@@ -98,13 +99,15 @@ impl ControllerCritical {
     }
 
     /// Set priority
-    pub fn set_priority(&self, channel_id: ChannelId, priority: super::Priority) {
+    pub fn set_priority(&self, channel_id: channel::Id, priority: super::Priority) {
         debug_assert_eq!(channel_id, 0);
-        self.dmac.chctrlb.write(|reg| reg.lvl().bits(priority as u8));
+        self.dmac
+            .chctrlb
+            .write(|reg| reg.lvl().bits(priority as u8));
     }
 
     /// Set interrupt mask
-    pub fn set_interrupt_mask(&self, channel_id: ChannelId, interrupt_mask: u8) {
+    pub fn set_interrupt_mask(&self, channel_id: channel::Id, interrupt_mask: u8) {
         debug_assert_eq!(channel_id, 0);
 
         unsafe {
@@ -123,7 +126,7 @@ impl ControllerCritical {
     }
 
     /// Set action
-    pub fn set_action(&self, channel_id: ChannelId, action: dmac::chctrlb::TRIGACTW) {
+    pub fn set_action(&self, channel_id: channel::Id, action: dmac::chctrlb::TRIGACTW) {
         unsafe {
             self.dmac.chid.write(|reg| reg.id().bits(channel_id));
             self.dmac.chctrlb.write(|reg| reg.trigact().variant(action));
@@ -131,7 +134,7 @@ impl ControllerCritical {
     }
 
     /// Set trigger
-    pub fn set_trigger(&self, channel_id: ChannelId, trigger: dmac::chctrlb::TRIGSRCW) {
+    pub fn set_trigger(&self, channel_id: channel::Id, trigger: dmac::chctrlb::TRIGSRCW) {
         unsafe {
             self.dmac.chid.write(|reg| reg.id().bits(channel_id));
             self.dmac
@@ -141,7 +144,7 @@ impl ControllerCritical {
     }
 
     /// Allocate the given channel
-    pub fn allocate_channel(&self, channel_id: ChannelId) -> Result<(), Error> {
+    pub fn allocate_channel(&self, channel_id: channel::Id) -> Result<(), Error> {
         let channel_mask = self.channel_mask();
 
         if channel_mask & (1 << channel_id) != 0 {
@@ -165,9 +168,9 @@ impl ControllerCritical {
     /// Release the given channel
     pub fn release_channel(
         &self,
-        channel_id: ChannelId,
+        channel_id: channel::Id,
         clock_source: &mut ClockSource,
-        nvic: &mut NVIC
+        nvic: &mut NVIC,
     ) -> Result<(), Error> {
         if self.channel_mask.get() & (1 << channel_id) != 0 {
             // Valid in-use channel; release it
@@ -194,7 +197,7 @@ impl ControllerCritical {
     }
 
     /// Abort DMA operation
-    pub fn abort(&self, channel_id: ChannelId) {
+    pub fn abort(&self, channel_id: channel::Id) {
         unsafe {
             // Select channel
             self.dmac.chid.write(|reg| reg.bits(channel_id));
